requirements.txt
pyserial==3.5
click==8.0.1
